tests: move quic tests into a separate file
